refactor: PlayerCommand 패키지 위치 view로 이동
